pyproject.toml
[project]
name = "gc_content_analyzer"
version = "0.1.0"
description = "A tool to analyze GC content and amino acid composition from genomic data"
authors = [
    { name = "Bruna", email = "your.email@example.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
dependencies = [
    "biopython",
    "pandas",
    "matplotlib",
    "seaborn"
]

[tool.setuptools.packages.find]
where = ["src"]


[project.urls]
Homepage = "https://github.com/BrunaGil25/gc-content-analyzer"
Repository = "https://github.com/BrunaGil25/gc-content-analyzer"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"